ChangeList: The _floodfill() graph library function has been fixed.
Note: The graph lib is abusing the pascal calling convention to
define custom calls. The graph lib broke because of changes to the
calling convention processing in the compiler, but I hesitate to call
that a bug. Ideally the graph lib shouldn't be abusing 'pascal' and
custom calling conventions should be defined in internal headers.

[git-p4: depot-paths = "//depot/openwatcom/": change = 34913]