ChangeList: The NT console utilities (vi, wd, wprof) now always force Quick Edit mode off. This makes the mouse functional again when Quick Edit is on by default, such as on Windows 10. Additionally, vi restores the console input mode on exit.
Implementor's Note: For reasons that are not obvious, SetConsoleMode
does not turn off Quick Edit mode unless the ENABLE_QUICK_EDIT_MODE flag
is clear *and* ENABLE_EXTENDED_FLAGS is set. This problem is old but was
not very obvious until Windows 10 enabled Quick Edit for console windows
by default.

[git-p4: depot-paths = "//depot/openwatcom/": change = 38198]